protosure/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.19.12
botocore==1.22.12
celery==5.2.7
certifi==2022.9.14
charset-normalizer==2.0.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.15
django-celery-beat==2.2.1
django-environ==0.4.5
django-extensions==3.1.5
django-filter==2.4.0
django-storages==1.12.3
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.6.0
idna==3.4
isort==5.9.3
Jinja2==3.0.3
jmespath==0.10.0
kombu==5.2.4
MarkupSafe==2.1.0
packaging==21.3
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.1.0
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
twilio==7.6.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
